headers: POSIX variadic fcntl on Linux; fixture: LLP64-safe shift widths
The Linux fcntl declaration's fixed int third parameter truncated
pointer arguments once calls narrow to the prototype (F_SETLK EFAULT
under the test prelude's include order; the fixture's own order let
unistd.h's correct variadic shape win). The shift-fold fixture used
long where the pattern needs 64 bits regardless of data model.

+++ tests/fixtures/c/sxtw_fold_source_liveness.c
@@ -1,10 +1,10 @@
 /* The Shl/Shr sign-narrow fold reads the Shl source's register at the
    Shr; an intervening definition may reuse that register once the
    source's live range ends. The fold must verify the place survives. */
-long f(long x, long b) {
-    long t = x << 32;
-    long v = x + b;
-    long y = t >> 32;
+long long f(long long x, long long b) {
+    long long t = x << 32;
+    long long v = x + b;
+    long long y = t >> 32;
     return y + v + b;
 }
